tree-optimization/107176 - SCEV analysis association issue
The following fixes a wrong-code issue caused by SCEV analysis
associating an addition due trying to use tail-recursion in
follow_ssa_edge_expr.  That causes us to apply a conversion at
the wrong point and thus miscompute the scalar evolution of
an induction variable.  This reverts the PR66375 fix and
revisits the PR42512 fix by keeping the evolution symbolic
up to the point we process the first linear function when
we then can check for the supported cases and substitute the
whole symbolic expression with the built chrec substituting
the proper initial value.

To simplify passing around things and to clarify scoping of
the involved functions this change wraps the SCEV DFS walking
code into a class.

Bootstrapped and tested on x86_64-unknown-linux-gnu, I'm doing
final testing on the following which includes some API
streamlining over the tested prototype.

	PR tree-optimization/107176
	PR tree-optimization/66375
	PR tree-optimization/42512
	* tree-scalar-evolution.cc (follow_ssa_edge_expr): Revert
	the PR66375 fix, do not not associate PLUS_EXPR to be able
	to use tail-recursion.
	(follow_ssa_edge_binary): Likewise.
	(interpret_loop_phi): Revert PR42512 fix, do not throw
	away analyze_evolution_in_loop result after the fact.
	(follow_ssa_edge_expr): When reaching halting_phi initalize
	the evolution to the symbolic value of the PHI result.
	(add_to_evolution_1): When adding the first evolution verify
	we can handle the expression wrapping the symbolic evolution
	and replace that in full using the initial condition.
	(class scev_dfs): New, contains ...
	(follow_ssa_edge_expr, follow_ssa_edge_binary,
	follow_ssa_edge_in_condition_phi_branch,
	follow_ssa_edge_in_condition_phi,
	follow_ssa_edge_inner_loop_phi,
	add_to_evolution, add_to_evolution_1): ... these with
	loop and halting_phi arguments in class data.
	(scev_dfs::get_ev): New toplevel DFS entry, start with
	a chrec_dont_know evolution.
	(analyze_evolution_in_loop): Use scev_dfs.

	* gcc.dg/torture/pr107176.c: New testcase.